refactor: remove unused SparkJobExecutionStrategy
1. remove unused SparkJobExecutionStrategy
